[#183] Refactored FileAccessor
This moves the class into the maps activity which solves the issues with not being able to call the appropriate android methods specific to an activity (in the map activity).
